refactor: move OutputFormat to separate settings.py
