+++ sage_acsv/__init__.py
@@ -1,2 +1,3 @@
 from sage_acsv.asymptotics import *
 from sage_acsv.kronecker import *
+from sage_acsv.settings import OutputFormat
